requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-text-splitters>=0.0.1
pypdf>=3.17.4
python-dotenv>=1.0.0
torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
--extra-index-url https://download.pytorch.org/whl/cpu
